refactor(gateway): extract instance tools and health handlers into submodules
Split gateway/tools.rs (was 1570 lines) by extracting instance lease
management into tools/instances.rs, and split
admin/application/handlers.rs (was 1594 lines) by extracting health and
reliability endpoints into application/health.rs. Both files now stay
below the 1500-line gate.
