+++ crates/dcc-mcp-gateway/src/gateway/admin/application/handlers.rs
@@ -25,13 +25,12 @@ use crate::gateway::admin::trace::{AgentContext, DispatchTrace};
 use crate::gateway::admin::update::{AdminInstanceUpdateVersion, admin_instance_update_version};
 use crate::gateway::capability::RefreshReason;
 use crate::gateway::capability_service::refresh_all_live_backends;
-use crate::gateway::resilience::{self as gw_resilience, gateway_limits};
 use crate::gateway::response_codec::{
     JSON_MIME, TOKEN_ESTIMATOR, TOON_MIME, default_rest_response_format,
 };
 use dcc_mcp_db::env::ENV_DCC_MCP_LOG_DIR;
 use dcc_mcp_db::read_gateway_log_dir_rows_recent;
-use dcc_mcp_transport::discovery::types::{GATEWAY_SENTINEL_DCC_TYPE, ServiceEntry};
+use dcc_mcp_transport::discovery::types::ServiceEntry;
 
 const ADMIN_FILE_LOG_READ_TIMEOUT: Duration = Duration::from_millis(750);
 
@@ -900,129 +899,6 @@ pub async fn handle_admin_logs(
     }))
 }
 
-/// `GET /admin/api/health` — service health summary.
-pub async fn handle_admin_health(State(s): State<AdminState>) -> impl IntoResponse {
-    let registry = s.gateway.registry.read().await;
-    let all = s.gateway.all_instances(&registry);
-    let ready = s.gateway.live_instances(&registry).len();
-    let gateway_sentinels = registry.list_instances(GATEWAY_SENTINEL_DCC_TYPE);
-    let total = all.len();
-    drop(registry);
-
-    let uptime_secs = s.started_at.elapsed().unwrap_or_default().as_secs();
-
-    let status = if ready > 0 || total == 0 {
-        "ok"
-    } else {
-        "degraded"
-    };
-
-    let limits = gateway_limits();
-    let circuits = gw_resilience::circuits().snapshot_json();
-    let rss_bytes = gateway_self_rss_bytes();
-
-    (
-        StatusCode::OK,
-        Json(json!({
-            "status": status,
-            "instances_ready": ready,
-            "instances_total": total,
-            "uptime_secs": uptime_secs,
-            "version": s.gateway.server_version,
-            "rss_bytes": rss_bytes,
-            "response_format": {
-                "default": default_rest_response_format().as_str(),
-                "legacy_mime": JSON_MIME,
-                "compact_mime": TOON_MIME,
-                "token_estimator": TOKEN_ESTIMATOR,
-            },
-            "gateway": gateway_health_snapshot(&gateway_sentinels),
-            "limits": {
-                "body_max_bytes": limits.body_max_bytes,
-                "rate_limit_per_minute_per_ip": limits.rate_limit_per_minute_per_ip,
-                "xff_trusted_depth": limits.xff_trusted_depth,
-                "read_retry_max": limits.read_retry_max,
-                "circuit_failure_threshold": limits.circuit_failure_threshold,
-                "circuit_open_secs": limits.circuit_open_secs,
-            },
-            "circuits": circuits,
-        })),
-    )
-}
-
-fn gateway_health_snapshot(sentinels: &[ServiceEntry]) -> Value {
-    let mut rows: Vec<Value> = sentinels.iter().map(gateway_sentinel_json).collect();
-    rows.sort_by(|a, b| {
-        let role_a = a.get("role").and_then(Value::as_str).unwrap_or("");
-        let role_b = b.get("role").and_then(Value::as_str).unwrap_or("");
-        let rank_a = if role_a == "active" { 0 } else { 1 };
-        let rank_b = if role_b == "active" { 0 } else { 1 };
-        rank_a.cmp(&rank_b).then_with(|| {
-            let ta = a
-                .get("last_heartbeat_unix")
-                .and_then(Value::as_u64)
-                .unwrap_or(0);
-            let tb = b
-                .get("last_heartbeat_unix")
-                .and_then(Value::as_u64)
-                .unwrap_or(0);
-            tb.cmp(&ta)
-        })
-    });
-    let current = rows
-        .iter()
-        .find(|row| row.get("role").and_then(Value::as_str) == Some("active"))
-        .cloned()
-        .or_else(|| rows.first().cloned());
-    let candidates: Vec<Value> = rows
-        .into_iter()
-        .filter(|row| row.get("role").and_then(Value::as_str) != Some("active"))
-        .collect();
-    json!({
-        "current": current,
-        "candidates": candidates,
-    })
-}
-
-fn gateway_sentinel_json(entry: &ServiceEntry) -> Value {
-    let last_heartbeat_secs = entry
-        .last_heartbeat
-        .duration_since(UNIX_EPOCH)
-        .ok()
-        .map(|d| d.as_secs());
-    let role = entry
-        .metadata
-        .get("gateway_role")
-        .cloned()
-        .unwrap_or_else(|| "active".to_string());
-    let name = entry
-        .metadata
-        .get("gateway_name")
-        .cloned()
-        .or_else(|| entry.display_name.clone())
-        .unwrap_or_else(|| format!("gateway-pid{}", entry.pid.unwrap_or_default()));
-    json!({
-        "name": name,
-        "role": role,
-        "pid": entry.pid,
-        "host": entry.host,
-        "port": entry.port,
-        "instance_id": entry.instance_id.to_string(),
-        "version": entry.version,
-        "adapter_version": entry.adapter_version,
-        "adapter_dcc": entry.adapter_dcc,
-        "last_heartbeat_unix": last_heartbeat_secs,
-        "metadata": entry.metadata,
-    })
-}
-
-fn gateway_self_rss_bytes() -> Option<u64> {
-    use sysinfo::{Pid, ProcessesToUpdate, System};
-    let mut sys = System::new();
-    let pid = Pid::from_u32(std::process::id());
-    sys.refresh_processes(ProcessesToUpdate::Some(std::slice::from_ref(&pid)), true);
-    sys.process(pid).map(|p| p.memory())
-}
 /// `GET /admin/api/traces?limit=200` — recent per-call dispatch traces (Phase 2).
 ///
 /// Each trace includes a waterfall of [`TraceSpan`]s plus optionally the
@@ -1471,124 +1347,4 @@ fn dispatch_trace_to_admin_row(t: &DispatchTrace, links: Option<AdminLinkBuilder
         row["links"] = links.request_links(&t.request_id);
     }
     row
-}
-
-/// `GET /admin/api/reliability` — reliability & stability summary (PIP-2766).
-///
-/// Aggregates health, circuit breaker state, capability funnel, and
-/// 24-hour stability (crashes, reconnects, recoveries) into a single
-/// payload for the admin UI Reliability panel.
-pub async fn handle_admin_reliability(State(s): State<AdminState>) -> impl IntoResponse {
-    let registry = s.gateway.registry.read().await;
-    let all = s.gateway.all_instances(&registry);
-    let ready = s.gateway.live_instances(&registry).len();
-    let total = all.len();
-    let gateway_sentinels = registry.list_instances(GATEWAY_SENTINEL_DCC_TYPE);
-    drop(registry);
-
-    let uptime_secs = s.started_at.elapsed().unwrap_or_default().as_secs();
-    let status = if ready > 0 || total == 0 { "ok" } else { "degraded" };
-    let limits = gateway_limits();
-    let circuits = gw_resilience::circuits().snapshot_json();
-    let rss_bytes = gateway_self_rss_bytes();
-
-    // Stability: query sessions table for crash/reconnect/recovery counts
-    // in the last 24 hours.
-    let (crashes_24h, reconnects_24h, recoveries_24h) = if let Some(ref lane) = s.admin_sqlite_lane
-    {
-        let reader = lane.reader();
-        let since_ms = std::time::SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_millis() as i64
-            - 86_400_000; // 24 hours
-
-        // Count sessions that ended abnormally in the last 24h
-        let all_sessions = reader.list_sessions(10_000, None, None);
-        let crashes = all_sessions
-            .iter()
-            .filter(|row| {
-                let status = row.get("status").and_then(|v| v.as_str()).unwrap_or("");
-                let ended = row.get("ended_at_ms").and_then(|v| v.as_i64()).unwrap_or(0);
-                (status == "crashed" || status == "gpu_crashed" || status == "disconnected")
-                    && ended >= since_ms
-            })
-            .count();
-
-        // Count session_events with event_type = 'reconnect' in the last 24h
-        let reconnects = all_sessions
-            .iter()
-            .filter(|row| {
-                let session_id = row.get("session_id").and_then(|v| v.as_str()).unwrap_or("");
-                let events = reader.list_session_events(session_id, 1000);
-                events.iter().any(|ev| {
-                    let event_type = ev.get("event_type").and_then(|v| v.as_str()).unwrap_or("");
-                    let created = ev.get("created_at_ms").and_then(|v| v.as_i64()).unwrap_or(0);
-                    event_type == "reconnect" && created >= since_ms
-                })
-            })
-            .count();
-
-        // Active sessions now (recoveries)
-        let recoveries = all_sessions
-            .iter()
-            .filter(|row| {
-                let status = row.get("status").and_then(|v| v.as_str()).unwrap_or("");
-                status == "active"
-            })
-            .count();
-
-        (crashes as i64, reconnects as i64, recoveries as i64)
-    } else {
-        (0, 0, 0)
-    };
-
-    // Stats for success rate — derive from existing stats if available,
-    // otherwise report 100% as safe default.
-    let success_rate = 100.0_f64;
-    let p50_latency_ms = 0_i64;
-
-    (
-        StatusCode::OK,
-        Json(json!({
-            "generated_at": chrono::Utc::now().to_rfc3339(),
-            "status": status,
-            "uptime_secs": uptime_secs,
-            "version": s.gateway.server_version,
-            "rss_bytes": rss_bytes,
-            "gateway": {
-                "status": status,
-                "uptime_secs": uptime_secs,
-                "version": s.gateway.server_version,
-                "election": gateway_health_snapshot(&gateway_sentinels),
-                "limits": {
-                    "body_max_bytes": limits.body_max_bytes,
-                    "rate_limit_per_minute_per_ip": limits.rate_limit_per_minute_per_ip,
-                    "circuit_failure_threshold": limits.circuit_failure_threshold,
-                    "circuit_open_secs": limits.circuit_open_secs,
-                },
-                "circuits": circuits,
-            },
-            "capability_funnel": {
-                "instances_ready": ready,
-                "instances_total": total,
-                "skills_loaded": 0,
-                "skills_total": 0,
-                "tools_registered": 0,
-                "resources_exposed": total,
-            },
-            "artifact_verification": {
-                "builds_verified": 0,
-                "builds_total": 0,
-                "verification_errors": 0,
-            },
-            "stability": {
-                "crashes_24h": crashes_24h,
-                "reconnects_24h": reconnects_24h,
-                "recoveries_24h": recoveries_24h,
-                "uptime_pct": success_rate,
-                "p50_latency_ms": p50_latency_ms,
-            },
-        })),
-    )
 }
